Type checking and unit tests pass, but there are still some files that need typing.
